backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
psycopg2-binary==2.9.10
pydantic==2.11.3
httpx==0.28.1
pytest==8.3.5
